Move from a build script to xtasks
A build script is great if the assets are guaranteed to be available,
but they would not be available once packaged and uploaded to a
registry.

Instead, this commit moves us to a system where `cargo xtask assets`
will generate the appropriate Rust file. This file can be checked in
immediately before a release is tagged and can then be a part of the
package.